feature/nlpi bot implementation (#13)
* initial code, WIP, NHLPI bot

* working code, current price done, get historical price WIP

* Update StatisticsNetherlandsPriceFeed.ts

This commit makes several changes to the StatisticsNetherlandsPriceFeed.ts file including adding a new function which increments the month returned by the response JSON by one so as to match the update time used by Statistics Netherlands. This ensure that the liquidation time provided is measured against the right time period as the response JSON current gives the date for the month that the price index represents, not the month that the price index was updated. Further changes are needed to be made to the getHistoricalPrice function as it is still failing some of my tests. I am hoping to have this resolved tonight.

* Update StatisticsNetherlandsPriceFeed.ts

This commit changes the logic of the getHistoricalPrice function such that the correct matchedIndex is returned based on the liquidation timestamp provided.

* Update StatisticsNetherlandsPriceFeed.ts

This commit adds a minor update to the getHistoricalPrice function logic

* updated code, to make findIndex code readable

---------

Co-authored-by: Kris O'Shea <minimallifter@protonmail.com>